timeutils: move cache related functions to a separate module
