+++ lib/template/macros.c
@@ -25,6 +25,7 @@
 #include "template/macros.h"
 #include "template/escaping.h"
 #include "timeutils/timeutils.h"
+#include "timeutils/cache.h"
 #include "logstamp.h"
 #include "messages.h"
 #include "str-format.h"
