+++ lib/timeutils/cache.h
@@ -1,6 +1,6 @@
 /*
- * Copyright (c) 2002-2014 Balabit
- * Copyright (c) 1998-2010 Balázs Scheidler
+ * Copyright (c) 2002-2018 Balabit
+ * Copyright (c) 1998-2018 Balázs Scheidler
  *
  * This library is free software; you can redistribute it and/or
  * modify it under the terms of the GNU Lesser General Public
@@ -22,27 +22,21 @@
  *
  */
 
-#ifndef TIMEUTILS_H_INCLUDED
-#define TIMEUTILS_H_INCLUDED
+#ifndef TIMEUTILS_CACHE_H_INCLUDED
+#define TIMEUTILS_CACHE_H_INCLUDED
 
 #include "syslog-ng.h"
 #include "compat/time.h"
 
-long get_local_timezone_ofs(time_t when);
+time_t cached_mktime(struct tm *tm);
+void cached_localtime(time_t *when, struct tm *tm);
+void cached_gmtime(time_t *when, struct tm *tm);
 
-gboolean check_nanosleep(void);
+void clean_time_cache(void);
 
-int format_zone_info(gchar *buf, size_t buflen, long gmtoff);
-glong g_time_val_diff(GTimeVal *t1, GTimeVal *t2);
-void timespec_add_msec(struct timespec *ts, glong msec);
-glong timespec_diff_msec(const struct timespec *t1, const struct timespec *t2);
-glong timespec_diff_nsec(struct timespec *t1, struct timespec *t2);
-gint determine_year_for_month(gint month, const struct tm *now);
-
-extern const char *month_names_abbrev[];
-extern const char *month_names[];
-extern const char *weekday_names_abbrev[];
-extern const char *weekday_names[];
-
+void invalidate_cached_time(void);
+void set_cached_time(GTimeVal *timeval);
+void cached_g_current_time(GTimeVal *result);
+time_t cached_g_current_time_sec(void);
 
 #endif
