+++ modules/dbparser/groupingby.c
@@ -28,6 +28,7 @@
 #include "str-utils.h"
 #include "filter/filter-expr.h"
 #include "timeutils/timeutils.h"
+#include "timeutils/cache.h"
 #include <iv.h>
 
 typedef struct _GroupingBy
