+++ modules/dbparser/patterndb.c
@@ -37,6 +37,7 @@
 #include "filter/filter-expr-parser.h"
 #include "logpipe.h"
 #include "timeutils/timeutils.h"
+#include "timeutils/cache.h"
 
 #include <string.h>
 #include <stdio.h>
