+++ modules/snmptrapd-parser/snmptrapd-header-parser.c
@@ -23,6 +23,7 @@
 
 #include "str-format.h"
 #include "timeutils/timeutils.h"
+#include "timeutils/cache.h"
 
 #include <string.h>
 #include <ctype.h>
