+++ tests/unit/test_dnscache.c
@@ -26,6 +26,7 @@
 #include "dnscache.h"
 #include "apphook.h"
 #include "timeutils/timeutils.h"
+#include "timeutils/cache.h"
 
 #include <sys/types.h>
 #include <sys/socket.h>
