+++ tests/unit/test_logqueue.c
@@ -31,6 +31,7 @@
 #include "mainloop-io-worker.h"
 #include "libtest/queue_utils_lib.h"
 #include "msg_parse_lib.h"
+#include "timeutils/timeutils.h"
 
 #include <stdlib.h>
 #include <string.h>
